Do not force the coalesced dequantize to params[0] dtype
The quantized coalesced path concatenates every parameter into one int8
buffer without grouping by dtype, unlike the non-quantized path which buckets
into one all-gather per dtype. Asking for params[0].dtype therefore casts the
whole flat buffer to whatever the first parameter happens to be, and the later
per-slice cast cannot undo rounding already applied to the rest.

Each slice is already cast to its own parameter's dtype when it is assigned,
so the flat buffer can keep the kernel's fp16 output as it did before.

+++ deepspeed/runtime/zero/partition_parameters.py
@@ -755,9 +755,11 @@ class AllGatherCoalescedHandle:
 
         if self.quantization:
             instrument_w_nvtx(self.quantization.quant_handle.wait)()
-            flat_tensor = self.quantization.backend.dequantize(self.quantization.quantized_param,
-                                                               self.quantization.scale_buffer,
-                                                               dtype=self.params[0].dtype).to(self.params[0].device)
+            # No dtype here on purpose. A quantized coalesced bucket is not grouped by dtype the
+            # way the non-quantized path is, so params[0].dtype is not necessarily the dtype of
+            # the rest of the bucket. Each slice is cast to its own parameter's dtype below.
+            flat_tensor = self.quantization.backend.dequantize(
+                self.quantization.quantized_param, self.quantization.scale_buffer).to(self.params[0].device)
 
             self.partitions: List[Parameter] = []
             for i in range(self.world_size):
